Build SleepRepository and test file
