+++ test/HydrationRepository-test.js
@@ -21,15 +21,15 @@ describe('UserRepository', function() {
   });
 
   it("should calculate the average daily water intake for all users", function() {
-    const avgDailyWater = userRepo.calculateAvgDailyWater();
+    const avgDailyWater = hydrationRepo.calculateAvgDailyWater();
 
     expect(avgDailyWater).to.equal(61);
   });
 
   it("should be able to retrieve the ounces drank by a user on a specific date", function() {
-    const numOunces1 = userRepo.retrieveNumOuncesByDate(1, '2019/06/18');
-    const numOunces2 = userRepo.retrieveNumOuncesByDate(2, '2019/06/16');
-    const numOunces3 = userRepo.retrieveNumOuncesByDate(3, '2019/06/20');
+    const numOunces1 = hydrationRepo.retrieveNumOuncesByDate(1, '2019/06/18');
+    const numOunces2 = hydrationRepo.retrieveNumOuncesByDate(2, '2019/06/16');
+    const numOunces3 = hydrationRepo.retrieveNumOuncesByDate(3, '2019/06/20');
 
     expect(numOunces1).to.equal(61);
     expect(numOunces2).to.equal(91);
@@ -37,9 +37,9 @@ describe('UserRepository', function() {
   });
 
   it("should calculate the average daily water intake for a user over the course of a week", function() {
-    const avgWeeklyWater1 = userRepo.calculateAvgWeeklyWater(1, '2019/06/15');
-    const avgWeeklyWater2 = userRepo.calculateAvgWeeklyWater(2, '2019/06/16');
-    const avgWeeklyWater3 = userRepo.calculateAvgWeeklyWater(3, '2019/06/17');
+    const avgWeeklyWater1 = hydrationRepo.calculateAvgWeeklyWater(1, '2019/06/15');
+    const avgWeeklyWater2 = hydrationRepo.calculateAvgWeeklyWater(2, '2019/06/16');
+    const avgWeeklyWater3 = hydrationRepo.calculateAvgWeeklyWater(3, '2019/06/17');
 
     expect(avgWeeklyWater1).to.equal(65);
     expect(avgWeeklyWater2).to.equal(70);
